+++ changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [1.12.2] -
+- add weekFirstDay, weekLastDay, weekAllDays
 - fix: handle +HH format for ISO8601 timeOffsetFormat etc.
 
 ## [1.12.1] - 2021-10-24